+++ src/Test/Models/TestEntry.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Hirtz\Cms\Test\Models;
 
 use Hirtz\Cms\Models\Entry;
+use Hirtz\Cms\Models\Types\EntryType;
 
 class TestEntry extends Entry
 {
@@ -15,14 +16,12 @@ class TestEntry extends Entry
     public static function getTypes(): array
     {
         return [
-            self::TYPE_PAGE => [
-                'name' => 'Page',
-                'hiddenFields' => ['content'],
-            ],
-            self::TYPE_POST => [
-                'name' => 'Post',
-                'hiddenFields' => ['#assets'],
-            ],
+            EntryType::make(self::TYPE_PAGE)
+                ->name('Page')
+                ->hiddenFields('content'),
+            EntryType::make(self::TYPE_POST)
+                ->name('Post')
+                ->hiddenFields(self::FIELD_ASSETS),
         ];
     }
 }